+++ packages/fictoan-react/src/components/Form/RadioButton/constants.ts
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { BaseInputComponentProps } from "../BaseInputComponent/constants";
+import { BaseInputComponentProps, ValueChangeHandler } from "../BaseInputComponent/constants";
 import { SpacingTypes } from "$components/Element/constants";
 
 // RADIO BUTTON ////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -8,7 +8,7 @@ export type RadioButtonElementType = HTMLDivElement;
 export type RadioButtonProps = Omit<BaseInputComponentProps<RadioButtonElementType>, "as" | "onChange" | "value"> & {
     value      : string; // Value is required for radio buttons
     checked  ? : boolean;
-    onChange ? : (value: string) => void;
+    onChange ? : ValueChangeHandler<string>;
 };
 
 // RADIO GROUP OPTIONS /////////////////////////////////////////////////////////////////////////////////////////////////
@@ -24,7 +24,7 @@ export interface RadioGroupCustomProps {
     options        : RadioGroupOptionProps[];
     value        ? : string;  // For controlled usage
     defaultValue ? : string;  // For uncontrolled usage
-    onChange     ? : (value: string) => void;  // Updated to handle direct values
+    onChange     ? : ValueChangeHandler<string>;
 }
 
 export type RadioGroupProps = Omit<RadioButtonProps, keyof RadioGroupCustomProps> & RadioGroupCustomProps & {